Agregado Feign Client en comments

+++ comments/src/main/java/sithmcfly/comments/service/ImpCommentService.java
@@ -1,7 +1,12 @@
 package sithmcfly.comments.service;
 
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.oauth2.jwt.Jwt;
 import org.springframework.stereotype.Service;
 import sithmcfly.comments.DTO.CommentDTO;
+import sithmcfly.comments.DTO.MovieDTO;
+import sithmcfly.comments.client.movieClient;
 import sithmcfly.comments.entities.Comment;
 import sithmcfly.comments.http.request.CommentCreateRequest;
 import sithmcfly.comments.http.request.CommentUpdateRequest;
@@ -17,9 +22,11 @@ import java.util.stream.Collectors;
 public class ImpCommentService implements ICommentService {
 
     private final CommentRepository commentRepository;
+    private final movieClient movieClient;
 
-    public ImpCommentService (CommentRepository commentRepository){
+    public ImpCommentService (CommentRepository commentRepository, movieClient movieClient){
         this.commentRepository = commentRepository;
+        this.movieClient = movieClient;
     }
 
     @Override
@@ -41,6 +48,7 @@ public class ImpCommentService implements ICommentService {
 
     @Override
     public CommentCreatedResponse createComment(CommentCreateRequest comment) {
+        MovieDTO movieExists = movieClient.getMovie(comment.getIdMovie());
         Comment createComment = CommentMapper.toEntity(comment);
 
         commentRepository.save(createComment);
@@ -51,6 +59,23 @@ public class ImpCommentService implements ICommentService {
     public void deleteComment(long id) {
         Comment comment = commentRepository.findById(id)
                 .orElseThrow(() -> new RuntimeException("no se encuentra"));
+        // Obtenemos el token
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        Jwt jwt = (Jwt) authentication.getPrincipal();
+        // Obtenemos el id del user
+        String idUserJwt = jwt.getSubject();
+        // Parseamos id que trae el token
+        long idUser = Long.parseLong(idUserJwt);
+        // Comprobamos si es ADMIN
+        boolean isAdmin = authentication.getAuthorities()
+                .stream()
+                .anyMatch(a -> a.getAuthority().equals("ROLE_ADMIN"));
+        // Comprobamos si es el propietario del comentario
+        boolean isOwner = comment.getIdUser() == idUser;
+        if(!isOwner && !isAdmin){
+            throw new RuntimeException("Fallo Forbidden");
+        }
+
         commentRepository.delete(comment);
     }
 
@@ -58,6 +83,22 @@ public class ImpCommentService implements ICommentService {
     public CommentUpdatedResponse editComment(long idComment, CommentUpdateRequest comment) {
         Comment updateComment = commentRepository.findById(idComment)
                 .orElseThrow(()-> new RuntimeException("no se encuentra"));
+        // Obtenemos el token
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        Jwt jwt = (Jwt) authentication.getPrincipal();
+        // Obtenemos el id del user
+        String idUserJwt = jwt.getSubject();
+        // Parseamos id que trae el token
+        long idUser = Long.parseLong(idUserJwt);
+        // Comprobamos si es ADMIN
+        boolean isAdmin = authentication.getAuthorities()
+                .stream()
+                .anyMatch(a -> a.getAuthority().equals("ROLE_ADMIN"));
+        // Comprobamos si es el propietario del comentario
+        boolean isOwner = updateComment.getIdUser() == idUser;
+        if(!isOwner && !isAdmin){
+            throw new RuntimeException("Fallo Forbidden");
+        }
         updateComment.setText(comment.getText());
 
         Comment saved = commentRepository.save(updateComment);
